requirements.txt
Django==3.1.7
PyMySQL==1.0.2
djangorestframework
django-cors-headers
